Se reestructura la logica del juego por archivos
