+++ game/index.html
@@ -6,7 +6,10 @@
     <title>Buscamina</title>
     <link rel="stylesheet" href="./style.css">
     <link rel="icon" href="../favicon_io/favicon.ico" type="image/x-icon">
-    <script src="./script.js" defer></script>
+    <!-- defer espera a que el html cargue completo antes de ejecutar el codigo js -->
+     <!-- module permite utilizar import y export en ES6 -->
+    <script type="module" src="./js/main.js"></script> 
+
 </head>
 <body>
     <div id="header">
