+++ game/js/levels.js
@@ -1,0 +1,5 @@
+export const  gameLevels= {
+  principiante: { rows: 8, columns: 8, mines: 10 },
+  intermedio:   { rows: 16, columns: 16, mines: 40 },
+  avanzado:     { rows: 30, columns: 40, mines: 99 }
+};
